fix: memory leak
By using DescriptorSetLayout outside VulkanEngine for materialLayout
