+++ src/include/graphics/vulkan/pipelines.h
@@ -2,14 +2,11 @@
 
 #include <filesystem>
 
-#include "vk_descriptors.h"
-#include "vk_images.h"
-#include "vk_initializers.h"
-#include "vk_pipelines.h"
-#include "vk_types.h"
-#include "IPipeline.h"
-#include "GraphicsPipeline.h"
 #include "ComputePipeline.h"
+#include "DescriptorSetLayout.h"
+#include "GraphicsPipeline.h"
+#include "vk_descriptors.h"
+#include "vk_types.h"
 
 class VulkanEngine;
 
@@ -17,7 +14,7 @@ struct GLTFMetallic_Roughness {
     MaterialPipeline opaquePipeline;
     MaterialPipeline transparentPipeline;
 
-    VkDescriptorSetLayout materialLayout;
+    DescriptorSetLayout materialLayout;
 
     struct MaterialConstants {
         glm::vec4 colorFactors;
